refactor(editor): extract templates and workflow types and apis
