+++ packages/frontend/@n8n/rest-api-client/src/api/index.ts
@@ -14,7 +14,9 @@ export * from './roles';
 export * from './settings';
 export * from './module-settings';
 export * from './sso';
+export * from './templates';
 export * from './ui';
 export * from './versions';
 export * from './webhooks';
 export * from './workflowHistory';
+export * from './workflows';
