+++ packages/frontend/editor-ui/src/components/WorkflowSelectorParameterInput/WorkflowSelectorParameterInput.vue
@@ -22,7 +22,7 @@ import { useProjectsStore } from '@/stores/projects.store';
 import { useTelemetry } from '@/composables/useTelemetry';
 import { VIEWS } from '@/constants';
 import { SAMPLE_SUBWORKFLOW_TRIGGER_ID, SAMPLE_SUBWORKFLOW_WORKFLOW } from '@/constants.workflows';
-import type { IWorkflowDataCreate } from '@/Interface';
+import type { WorkflowDataCreate } from '@n8n/rest-api-client/api/workflows';
 import { useDocumentVisibility } from '@/composables/useDocumentVisibility';
 
 export interface Props {
@@ -36,7 +36,7 @@ export interface Props {
 	forceShowExpression?: boolean;
 	parameterIssues?: string[];
 	parameter: INodeProperties;
-	sampleWorkflow?: IWorkflowDataCreate;
+	sampleWorkflow?: WorkflowDataCreate;
 	newResourceLabel?: string;
 }
 
@@ -254,7 +254,7 @@ const onAddResourceClicked = async () => {
 		(w) => w.name && new RegExp(workflowName).test(w.name),
 	);
 
-	const workflow: IWorkflowDataCreate = {
+	const workflow: WorkflowDataCreate = {
 		...sampleWorkflow,
 		name: `${workflowName} ${sampleSubWorkflows.length + 1}`,
 	};
